move sign clip processors; SignCLIPMetaProcessorV2 sequential access items; try perceiver
