+++ examples/MMPT/projects/retri/signclip_bsl/bobsl_islr_perceiver.yaml
@@ -1,0 +1,57 @@
+dataset:
+  debug: true
+  max_video_len: 32
+  max_len: 32
+  meta_processor: SignCLIPMetaProcessorV2
+  video_processor: SignCLIPPoseProcessor
+  text_processor: TextProcessor
+  bert_name: bert-base-cased
+  data_dir: /scratch/shared/beegfs/zifan/tensorflow_datasets
+  train_datasets: [['bobsl_islr', '1.4.0']]
+  valid_datasets: [['bobsl_islr', '1.4.0']]
+  test_path: data/<dataset>/test.txt # placeholder, no use
+  val_path: data/<dataset>/val.txt # placeholder, no use
+  train_path: data/<dataset>/train.txt # placeholder, no use
+  aligner: DSAligner
+  pose_components: 'reduced_face'
+  sp_universal_tagging: true
+fairseq:
+  common:
+    fp16: true
+    tensorboard_logdir: run
+    log_interval: 1000
+  dataset:
+    num_workers: 0
+    # batch_size: 1024
+    batch_size: 32
+  optimization:
+    lr:
+    - 5.0e-05
+    clip_norm: 2.0
+    optimizer: adam
+    adam_betas: (0.9, 0.98)
+    lr_scheduler: polynomial_decay
+    total_num_update: 1000000
+    warmup_updates: 122
+    weight_decay: 0.0
+    ddp_backend: no_c10d
+    max_epoch: 25
+  checkpoint:
+    # restore_file: runs/retri_v1_1/baseline_temporal/checkpoint_best.pt
+    no_epoch_checkpoints: true
+    reset_optimizer: true
+    reset_dataloader: true
+    reset_meters: true
+    save_dir: runs/retri_bsl/bobsl_islr_perceiver
+task_type: sweep_small
+model:
+  model_cls: MMFusionSeparate
+  mm_encoder_cls: null
+  video_encoder_cls: MMBertForEncoder
+  text_encoder_cls: BertModel
+  num_hidden_video_layers: 6
+  vfeat_dim: 609
+loss:
+  loss_cls: MMContraLoss
+eval:
+  save_path: runs/retri_bsl/bobsl_islr_perceiver
